Shrink Terminal interface to necessary subset for non-Console users, keep platform specific bits platform specific
