+++ src/platform/CMakeLists.txt
@@ -27,4 +27,3 @@ add_subdirectory(backends)
 add_subdirectory(client)
 add_subdirectory(console)
 add_subdirectory(logger)
-add_subdirectory(terminal)
