+++ src/platform/console/console.cpp
@@ -19,12 +19,13 @@
 #include <multipass/terminal.h>
 
 #include "unix_console.h"
+#include "unix_terminal.h"
 
 namespace mp = multipass;
 
 mp::Console::UPtr mp::Console::make_console(ssh_channel channel, Terminal* term)
 {
-    return std::make_unique<UnixConsole>(channel, term);
+    return std::make_unique<UnixConsole>(channel, static_cast<UnixTerminal*>(term));
 }
 
 void mp::Console::setup_environment()
